backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==4.2.11
django-cors-headers==3.13.0
django-filter==24.2
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-extra-fields==3.7.0
exceptiongroup==1.2.1
filetype==1.2.0
flake8==6.0.0
gunicorn==20.1.0
idna==3.7
iniconfig==2.0.0
isort==5.13.2
MarkupSafe==2.1.5
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
psycopg2-binary==2.9.3
pycodestyle==2.10.0
pycparser==2.22
pyflakes==3.0.1
PyJWT==2.8.0
pyperclip==1.8.2
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
